move localized_string_id to own file
I dont know if there was a circular dependency issue
