+++ rpcs3/Emu/localized_string_id.h
@@ -1,0 +1,115 @@
+﻿#pragma once
+
+enum class localized_string_id
+{
+	INVALID,
+
+	RSX_OVERLAYS_TROPHY_BRONZE,
+	RSX_OVERLAYS_TROPHY_SILVER,
+	RSX_OVERLAYS_TROPHY_GOLD,
+	RSX_OVERLAYS_TROPHY_PLATINUM,
+	RSX_OVERLAYS_COMPILING_SHADERS,
+	RSX_OVERLAYS_MSG_DIALOG_YES,
+	RSX_OVERLAYS_MSG_DIALOG_NO,
+	RSX_OVERLAYS_MSG_DIALOG_CANCEL,
+	RSX_OVERLAYS_MSG_DIALOG_OK,
+	RSX_OVERLAYS_SAVE_DIALOG_TITLE,
+	RSX_OVERLAYS_SAVE_DIALOG_DELETE,
+	RSX_OVERLAYS_SAVE_DIALOG_LOAD,
+	RSX_OVERLAYS_SAVE_DIALOG_SAVE,
+	RSX_OVERLAYS_OSK_DIALOG_ACCEPT,
+	RSX_OVERLAYS_OSK_DIALOG_CANCEL,
+	RSX_OVERLAYS_OSK_DIALOG_SPACE,
+	RSX_OVERLAYS_OSK_DIALOG_BACKSPACE,
+	RSX_OVERLAYS_OSK_DIALOG_SHIFT,
+	RSX_OVERLAYS_OSK_DIALOG_ENTER_TEXT,
+	RSX_OVERLAYS_OSK_DIALOG_ENTER_PASSWORD,
+	RSX_OVERLAYS_LIST_SELECT,
+	RSX_OVERLAYS_LIST_CANCEL,
+
+	CELL_GAME_ERROR_BROKEN_GAMEDATA,
+	CELL_GAME_ERROR_BROKEN_HDDGAME,
+	CELL_GAME_ERROR_BROKEN_EXIT_GAMEDATA,
+	CELL_GAME_ERROR_BROKEN_EXIT_HDDGAME,
+	CELL_GAME_ERROR_NOSPACE,
+	CELL_GAME_ERROR_NOSPACE_EXIT,
+	CELL_GAME_ERROR_DIR_NAME,
+	CELL_GAME_DATA_EXIT_BROKEN,
+	CELL_HDD_GAME_EXIT_BROKEN,
+
+	CELL_MSG_DIALOG_ERROR_DEFAULT,
+	CELL_MSG_DIALOG_ERROR_80010001,
+	CELL_MSG_DIALOG_ERROR_80010002,
+	CELL_MSG_DIALOG_ERROR_80010003,
+	CELL_MSG_DIALOG_ERROR_80010004,
+	CELL_MSG_DIALOG_ERROR_80010005,
+	CELL_MSG_DIALOG_ERROR_80010006,
+	CELL_MSG_DIALOG_ERROR_80010007,
+	CELL_MSG_DIALOG_ERROR_80010008,
+	CELL_MSG_DIALOG_ERROR_80010009,
+	CELL_MSG_DIALOG_ERROR_8001000A,
+	CELL_MSG_DIALOG_ERROR_8001000B,
+	CELL_MSG_DIALOG_ERROR_8001000C,
+	CELL_MSG_DIALOG_ERROR_8001000D,
+	CELL_MSG_DIALOG_ERROR_8001000F,
+	CELL_MSG_DIALOG_ERROR_80010010,
+	CELL_MSG_DIALOG_ERROR_80010011,
+	CELL_MSG_DIALOG_ERROR_80010012,
+	CELL_MSG_DIALOG_ERROR_80010013,
+	CELL_MSG_DIALOG_ERROR_80010014,
+	CELL_MSG_DIALOG_ERROR_80010015,
+	CELL_MSG_DIALOG_ERROR_80010016,
+	CELL_MSG_DIALOG_ERROR_80010017,
+	CELL_MSG_DIALOG_ERROR_80010018,
+	CELL_MSG_DIALOG_ERROR_80010019,
+	CELL_MSG_DIALOG_ERROR_8001001A,
+	CELL_MSG_DIALOG_ERROR_8001001B,
+	CELL_MSG_DIALOG_ERROR_8001001C,
+	CELL_MSG_DIALOG_ERROR_8001001D,
+	CELL_MSG_DIALOG_ERROR_8001001E,
+	CELL_MSG_DIALOG_ERROR_8001001F,
+	CELL_MSG_DIALOG_ERROR_80010020,
+	CELL_MSG_DIALOG_ERROR_80010021,
+	CELL_MSG_DIALOG_ERROR_80010022,
+	CELL_MSG_DIALOG_ERROR_80010023,
+	CELL_MSG_DIALOG_ERROR_80010024,
+	CELL_MSG_DIALOG_ERROR_80010025,
+	CELL_MSG_DIALOG_ERROR_80010026,
+	CELL_MSG_DIALOG_ERROR_80010027,
+	CELL_MSG_DIALOG_ERROR_80010028,
+	CELL_MSG_DIALOG_ERROR_80010029,
+	CELL_MSG_DIALOG_ERROR_8001002A,
+	CELL_MSG_DIALOG_ERROR_8001002B,
+	CELL_MSG_DIALOG_ERROR_8001002C,
+	CELL_MSG_DIALOG_ERROR_8001002D,
+	CELL_MSG_DIALOG_ERROR_8001002E,
+	CELL_MSG_DIALOG_ERROR_8001002F,
+	CELL_MSG_DIALOG_ERROR_80010030,
+	CELL_MSG_DIALOG_ERROR_80010031,
+	CELL_MSG_DIALOG_ERROR_80010032,
+	CELL_MSG_DIALOG_ERROR_80010033,
+	CELL_MSG_DIALOG_ERROR_80010034,
+	CELL_MSG_DIALOG_ERROR_80010035,
+	CELL_MSG_DIALOG_ERROR_80010036,
+	CELL_MSG_DIALOG_ERROR_80010037,
+	CELL_MSG_DIALOG_ERROR_80010038,
+	CELL_MSG_DIALOG_ERROR_80010039,
+	CELL_MSG_DIALOG_ERROR_8001003A,
+	CELL_MSG_DIALOG_ERROR_8001003B,
+	CELL_MSG_DIALOG_ERROR_8001003C,
+	CELL_MSG_DIALOG_ERROR_8001003D,
+	CELL_MSG_DIALOG_ERROR_8001003E,
+
+	CELL_OSK_DIALOG_TITLE,
+
+	CELL_SAVEDATA_CB_BROKEN,
+	CELL_SAVEDATA_CB_FAILURE,
+	CELL_SAVEDATA_CB_NO_DATA,
+	CELL_SAVEDATA_CB_NO_SPACE,
+	CELL_SAVEDATA_CREATE_CONFIRMATION,
+	CELL_SAVEDATA_DELETE_CONFIRMATION,
+	CELL_SAVEDATA_DELETE_SUCCESS,
+	CELL_SAVEDATA_DELETE,
+	CELL_SAVEDATA_LOAD,
+	CELL_SAVEDATA_OVERWRITE,
+};
